requirements.txt
arcadepy
openai
python-dotenv
pydantic
flask>=3.0.0
flask-cors>=4.0.0